third_party/WebKit/LayoutTests/platform/win/virtual/mojo-loading/http/tests/security/contentSecurityPolicy/nonces/script-multiple-blocked-expected.txt

CONSOLE ERROR: The Content Security Policy 'script-src 'unsafe-inline' 'self'' was delivered in report-only mode, but does not specify a 'report-uri'; the policy will have no effect. Please either add a 'report-uri' directive, or deliver the policy via the 'Content-Security-Policy' header.
CONSOLE ERROR: line 23: Refused to execute inline script because it violates the following Content Security Policy directive: "script-src 'self' 'nonce-abc'". Either the 'unsafe-inline' keyword, a hash ('sha256-eEPInijSMac6HLcSDJI0ovguR7rqdEbs55HMFzqd82Y='), or a nonce ('nonce-...') is required to enable inline execution.

CONSOLE ERROR: line 26: Refused to execute inline script because it violates the following Content Security Policy directive: "script-src 'self' 'nonce-abc'". Either the 'unsafe-inline' keyword, a hash ('sha256-eEPInijSMac6HLcSDJI0ovguR7rqdEbs55HMFzqd82Y='), or a nonce ('nonce-...') is required to enable inline execution.

This is a testharness.js-based test.
PASS Unnonced script blocks generate reports. 
PASS Nonced script executes, and does not generate a violation report. 
PASS Blocks without correct nonce do not execute, and generate violation reports 
Harness: the test ran to completion.
